Move leaderboard logic into leaderboard.go
